backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["skorikovdo <skorikov.denis.o@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
uvicorn = {extras = ["standard"], version = "^0.30.3"}
gunicorn = "^22.0.0"
psycopg2 = "^2.9.9"
pydantic-settings = "^2.3.4"
pyjwt = "^2.9.0"
sqlalchemy = "^2.0.31"
bcrypt = "^4.2.0"
alembic = "^1.13.2"

[tool.poetry.group.dev.dependencies]
httpie = "^3.2.3"
flake8 = "^7.1.0"
black = "^24.4.2"
mypy = "^1.11.0"
pylint = "^3.2.6"
isort = "^5.13.2"
sqladmin = "^0.18.0"
itsdangerous = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 79
